+++ src/simple/createPub.ts
@@ -76,8 +76,6 @@ else
     log_error "💥 发布 😔 失败 💔 " 
     exit 1
 fi
-
-
 `,
   );
 }