+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # 日志
 
+!!! 更新依赖后需注意在 [buildDevDependencies](./src/data-store/index.ts) 方法更新返回值
+
+## 1.3.0 (2026-7-1)
+
+- 修改 `tsconfig.types.json` 导出文件（ declarationDir ）路径及根（ rootDir ）
+- 移除  LICENSE 文件中的中文部分（冗余且显得和 low ）
+- 移除 `.prettierignore` 文件
+
 ## 1.2.4 (2026-7-1)
 
 - 修复已知问题